Move usageTrackers to relevant hooks
